backend/requirements.txt
appdirs==1.4.3
argon2-cffi==19.2.0
astroid==2.4.0
attrs==19.3.0
bcrypt==3.1.7
black==19.10b0
cffi==1.14.0
click==7.1.2
colorama==0.4.3
databases==0.3.2
dnspython==1.16.0
email-validator==1.0.5
fastapi==0.54.1
h11==0.9.0
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
passlib==1.7.2
pathspec==0.8.0
pycparser==2.20
pydantic==1.5.1
PyJWT==1.7.1
pylint==2.5.0
pylint-django==2.0.15
pylint-plugin-utils==0.6
python-multipart==0.0.5
regex==2020.4.4
six==1.14.0
SQLAlchemy==1.3.16
starlette==0.13.2
toml==0.10.0
typed-ast==1.4.1
uvicorn==0.11.5
websockets==8.1
wrapt==1.12.1
